+++ releases/ChangeLog.md
@@ -1,5 +1,6 @@
 ## 5.2.3 - 2024-XX-XX
 
+- Enhancement: Allow to specify start index for address explorer export and browsing
 - Change: `Passphrase` menu item is no longer offered if BIP39 passphrase
   already in use. Use `Restore Master` with ability to keep or purge current
   passphrase wallet settings.